Change byte to float

+++ train-gray-2.lua
@@ -9,6 +9,6 @@ local file_name = file_name_route .. '00000' .. tostring(file_num) .. '.jpg'
 
 print(file_name)
 
-local image_input = image.load(file_name, 3, 'byte')
+local image_input = image.load(file_name, 3, 'float')
 
 print(image_input)
